[HWKMETRICS-72] move schema manager code into its own module
