+++ schema-manager/src/main/java/org/hawkular/metrics/schema/TokenReplacingReader.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.core.impl.util;
+package org.hawkular.metrics.schema;
 
 import java.io.IOException;
 import java.io.PushbackReader;
